Enhance resource allocation flexibility and update deployment guidelines
- Updated resource allocation parameters in update_cluster_tool.py, update_devbox_tool.py, and update_launchpad_tool.py to accept fractional values (0.5) for CPU and memory, improving flexibility in resource management.
- Revised related data models in update.py files to reflect changes in CPU and memory types, ensuring consistency across the codebase.
- Clarified deployment guidelines in deploy_project_prompt.py to emphasize the limitation of deploying only one type at a time, enhancing user understanding of deployment processes.

+++ src/lib/brain/sealos/cluster/update.py
@@ -26,8 +26,8 @@ class ClusterUpdateData(BaseModel):
     name: str = Field(..., description="Cluster name")
     # Add fields based on clusterUpdateFormSchema from the Next.js code
     # This is a placeholder - you may need to adjust based on actual schema
-    cpu: Optional[int] = Field(None, description="CPU allocation")
-    memory: Optional[int] = Field(None, description="Memory allocation")
+    cpu: Optional[float] = Field(None, description="CPU allocation")
+    memory: Optional[float] = Field(None, description="Memory allocation")
     replicas: Optional[int] = Field(None, description="Number of replicas")
     storage: Optional[int] = Field(None, description="Storage allocation in GB")
     # Add other fields as needed based on the actual clusterUpdateFormSchema
