+++ src/lib/brain/sealos/devbox/update.py
@@ -26,8 +26,8 @@ class DevboxUpdateData(BaseModel):
     name: str = Field(..., description="Devbox name")
     # Add fields based on devboxUpdateFormSchema from the Next.js code
     # This is a placeholder - you may need to adjust based on actual schema
-    cpu: Optional[int] = Field(None, description="CPU allocation")
-    memory: Optional[int] = Field(None, description="Memory allocation")
+    cpu: Optional[float] = Field(None, description="CPU allocation")
+    memory: Optional[float] = Field(None, description="Memory allocation")
     createPorts: Optional[List[int]] = Field(
         None, description="Array of port numbers to create"
     )
